Include metadata in cut index
Instead of listing only the cut IDs, the create_index script now
includes the cut metadata (contents of the cut file minus the "slicers"
property), and adds an "href" property with a (potentially relative)
URL to the actual cut file.

The JavaScript for the index page was adapted to use the info from
that index instead of downloading all cut files.

The one-liner generation was adapted to correctly apply the cut's href
to the document's base URI, and also to use the cut's ID (which is
presumed unique and filename-safe) to name the uploaded JSON file,
since we no longer assume the cut has a "filename", just an URL.